requirements.txt
asgiref==3.11.1
attr==0.3.2
colorama==0.4.6
ConfigParser==7.2.0
cryptography==46.0.5
docutils==0.22.4
filelock==3.25.2
flask_sqlalchemy==3.1.1
HTMLParser==0.0.2
importlib_metadata==8.7.1
ipython==8.12.3
ipywidgets==8.1.8
jnius==1.1.0
keyring==25.7.0
protobuf==7.34.1
pyOpenSSL==26.0.0
python-dotenv==1.2.2
redis==7.4.0
Sphinx==9.1.0
thread==2.0.6
typing_extensions==4.15.0
urllib3_secure_extra==0.1.0
watchdog==6.0.0
xmlrpclib==1.0.1
